apps/api/requirements.txt
beautifulsoup4==4.13.4
fastapi==0.115.12
httpx==0.28.1
jose==1.0.0
python-dotenv==1.1.0
python_jose==3.4.0
pytz==2025.2
redis==6.1.0
stripe==12.1.0
uvicorn==0.34.2
jinja2
python-multipart
pydantic[email]
pydantic-settings==2.13.1
sqlalchemy
pymysql
bcrypt
uvicorn[standard]
motor
certifi
colorama
openai
